Lights for intake go longer and stop when intook
Added a few functions in lights to name whatever effect the scheduler is using, so checks like when the command ends is easy, as you can tell if your actually stopping the light effect you want to stop.

As far as functionality goes, I basically extended the time the lights will flash when intaking a game piece, and immediately stop the effect when a piece is intaken.